Revise TestEngine default implementations
Addresses #1114
